+++ src/lib.rs
@@ -100,7 +100,7 @@ pub trait Filesystem {
     /// each forget. The filesystem may ignore forget calls, if the inodes don't need to
     /// have a limited lifetime. On unmount it is not guaranteed, that all referenced
     /// inodes will receive a forget message.
-    fn forget (&mut self, _req: &Request, _ino: u64, _nlookup: uint) {
+    fn forget (&mut self, _req: &Request, _ino: u64, _nlookup: usize) {
     }
 
     /// Get file attributes
@@ -162,7 +162,7 @@ pub trait Filesystem {
     /// anything in fh. There are also some flags (direct_io, keep_cache) which the
     /// filesystem may set, to change the way the file is opened. See fuse_file_info
     /// structure in <fuse_common.h> for more details.
-    fn open (&mut self, _req: &Request, _ino: u64, _flags: uint, reply: ReplyOpen) {
+    fn open (&mut self, _req: &Request, _ino: u64, _flags: usize, reply: ReplyOpen) {
         reply.opened(0, 0);
     }
 
@@ -173,7 +173,7 @@ pub trait Filesystem {
     /// return value of the read system call will reflect the return value of this
     /// operation. fh will contain the value set by the open method, or will be undefined
     /// if the open method didn't set any value.
-    fn read (&mut self, _req: &Request, _ino: u64, _fh: u64, _offset: u64, _size: uint, reply: ReplyData) {
+    fn read (&mut self, _req: &Request, _ino: u64, _fh: u64, _offset: u64, _size: usize, reply: ReplyData) {
         reply.error(ENOSYS);
     }
 
@@ -183,7 +183,7 @@ pub trait Filesystem {
     /// which case the return value of the write system call will reflect the return
     /// value of this operation. fh will contain the value set by the open method, or
     /// will be undefined if the open method didn't set any value.
-    fn write (&mut self, _req: &Request, _ino: u64, _fh: u64, _offset: u64, _data: &[u8], _flags: uint, reply: ReplyWrite) {
+    fn write (&mut self, _req: &Request, _ino: u64, _fh: u64, _offset: u64, _data: &[u8], _flags: usize, reply: ReplyWrite) {
         reply.error(ENOSYS);
     }
 
@@ -209,7 +209,7 @@ pub trait Filesystem {
     /// the release. fh will contain the value set by the open method, or will be undefined
     /// if the open method didn't set any value. flags will contain the same flags as for
     /// open.
-    fn release (&mut self, _req: &Request, _ino: u64, _fh: u64, _flags: uint, _lock_owner: u64, _flush: bool, reply: ReplyEmpty) {
+    fn release (&mut self, _req: &Request, _ino: u64, _fh: u64, _flags: usize, _lock_owner: u64, _flush: bool, reply: ReplyEmpty) {
         reply.ok();
     }
 
@@ -227,7 +227,7 @@ pub trait Filesystem {
     /// anything in fh, though that makes it impossible to implement standard conforming
     /// directory stream operations in case the contents of the directory can change
     /// between opendir and releasedir.
-    fn opendir (&mut self, _req: &Request, _ino: u64, _flags: uint, reply: ReplyOpen) {
+    fn opendir (&mut self, _req: &Request, _ino: u64, _flags: usize, reply: ReplyOpen) {
         reply.opened(0, 0);
     }
 
@@ -244,7 +244,7 @@ pub trait Filesystem {
     /// For every opendir call there will be exactly one releasedir call. fh will
     /// contain the value set by the opendir method, or will be undefined if the
     /// opendir method didn't set any value.
-    fn releasedir (&mut self, _req: &Request, _ino: u64, _fh: u64, _flags: uint, reply: ReplyEmpty) {
+    fn releasedir (&mut self, _req: &Request, _ino: u64, _fh: u64, _flags: usize, reply: ReplyEmpty) {
         reply.ok();
     }
 
@@ -262,7 +262,7 @@ pub trait Filesystem {
     }
 
     /// Set an extended attribute
-    fn setxattr (&mut self, _req: &Request, _ino: u64, _name: &[u8], _value: &[u8], _flags: uint, _position: u32, reply: ReplyEmpty) {
+    fn setxattr (&mut self, _req: &Request, _ino: u64, _name: &[u8], _value: &[u8], _flags: usize, _position: u32, reply: ReplyEmpty) {
         reply.error(ENOSYS);
     }
 
@@ -289,7 +289,7 @@ pub trait Filesystem {
     /// This will be called for the access() system call. If the 'default_permissions'
     /// mount option is given, this method is not called. This method is not called
     /// under Linux kernel versions 2.4.x
-    fn access (&mut self, _req: &Request, _ino: u64, _mask: uint, reply: ReplyEmpty) {
+    fn access (&mut self, _req: &Request, _ino: u64, _mask: usize, reply: ReplyEmpty) {
         reply.error(ENOSYS);
     }
 
@@ -303,7 +303,7 @@ pub trait Filesystem {
     /// structure in <fuse_common.h> for more details. If this method is not
     /// implemented or under Linux kernel versions earlier than 2.6.15, the mknod()
     /// and open() methods will be called instead.
-    fn create (&mut self, _req: &Request, _parent: u64, _name: &PosixPath, _mode: u32, _flags: uint, reply: ReplyCreate) {
+    fn create (&mut self, _req: &Request, _parent: u64, _name: &PosixPath, _mode: u32, _flags: usize, reply: ReplyCreate) {
         reply.error(ENOSYS);
     }
 
@@ -326,7 +326,7 @@ pub trait Filesystem {
     /// Map block index within file to block index within device
     /// Note: This makes sense only for block device backed filesystems mounted
     /// with the 'blkdev' option
-    fn bmap (&mut self, _req: &Request, _ino: u64, _blocksize: uint, _idx: u64, reply: ReplyBmap) {
+    fn bmap (&mut self, _req: &Request, _ino: u64, _blocksize: usize, _idx: u64, reply: ReplyBmap) {
         reply.error(ENOSYS);
     }
 
@@ -339,7 +339,7 @@ pub trait Filesystem {
 
     /// OS X only (undocumented)
     #[cfg(target_os = "macos")]
-    fn exchange (&mut self, _req: &Request, _parent: u64, _name: &PosixPath, _newparent: u64, _newname: &PosixPath, _options: uint, reply: ReplyEmpty) {
+    fn exchange (&mut self, _req: &Request, _parent: u64, _name: &PosixPath, _newparent: u64, _newname: &PosixPath, _options: usize, reply: ReplyEmpty) {
         reply.error(ENOSYS);
     }
